+++ src/routes/ManageAds/AdsHeader/AdsCreateButton/adsCreateButton.module.scss
@@ -18,7 +18,7 @@
   right: 0;
   z-index: levels.$DROPDOWN_BOTTOM;
   width: 300px;
-  padding: 10px 20px 5px;
+  padding: 10px 20px 20px;
   margin-top: 4px;
   background-color: colors.$WHITE;
   border: 1px solid colors.$GRAY100;
