+++ src/routes/ManageAds/AdsHeader/AdsCreateForm/adsCreateForm.module.scss
@@ -1,0 +1,26 @@
+@use 'styles/constants/colors';
+@use 'styles/mixins/flexbox';
+
+.createFormWrapper {
+  @include flexbox.flexbox(center, center);
+  flex-direction: column;
+
+  label {
+    width: 100px;
+    margin-right: 10px;
+    font-size: 12px;
+    color: colors.$GRAY100;
+  }
+
+  ul {
+    width: 100%;
+    margin: 20px 0;
+  }
+
+  button {
+    width: 100px;
+    height: 40px;
+    border: 1px solid colors.$GRAY100;
+    border-radius: 10px;
+  }
+}
