--- src/routes/ManageAds/AdsHeader/adsCreateForm.module.scss
@@ -1,65 +1,0 @@
-@use 'styles/constants/colors';
-@use 'styles/mixins/flexbox';
-
-.createFormWrapper {
-  @include flexbox.flexbox(center, center);
-  flex-direction: column;
-
-  label {
-    width: 100px;
-    margin-right: 10px;
-    font-size: 12px;
-    color: colors.$GRAY100;
-  }
-
-  ul {
-    width: 100%;
-  }
-}
-
-.createList {
-  @include flexbox.flexbox(start, center);
-  flex-direction: row;
-  width: 100%;
-  height: 40px;
-  border-top: 1px solid colors.$GRAY50;
-
-  &:last-child {
-    border-bottom: 1px solid colors.$GRAY50;
-  }
-
-  input[type='radio'],
-  input[type='checkbox'] {
-    width: 20px;
-    height: 20px;
-    border: 2px solid colors.$GRAY100;
-    border-radius: 50%;
-
-    &:checked {
-      background-color: colors.$GRAY100;
-    }
-  }
-
-  input[type='date'],
-  input[type='number'] {
-    width: 120px;
-    color: colors.$GRAY300;
-  }
-}
-
-.radioInput {
-  @include flexbox.flexbox(start, center);
-  flex-direction: row;
-  width: 140px;
-
-  label {
-    width: 40px;
-    margin: 0 5px;
-    line-height: 20px;
-    color: colors.$GRAY300;
-  }
-}
-
-.isInputHidden {
-  display: none;
-}
